+++ src/tools/UpdatePackageDependencies.ts
@@ -5,8 +5,7 @@
 
 import * as fs from 'fs';
 import * as os from 'os';
-import { PackageJSONPackage } from '../packageManager/Package';
-
+import { PackageJSONPackage } from '../packageManager/PackageJSONPackage';
 
 interface PackageJSONFile
 {
