+++ test/featureTests/OmnisharpPackageCreator.test.ts
@@ -6,7 +6,7 @@
 import { assert, should, expect } from "chai";
 import { SetBinaryAndGetPackage, GetPackagesFromVersion } from "../../src/omnisharp/OmnisharpPackageCreator";
 import { testPackageJSON } from "./testAssets/testAssets";
-import { PackageJSONPackage } from "../../src/packageManager/Package";
+import { PackageJSONPackage } from "../../src/packageManager/PackageJSONPackage";
 
 suite("GetOmnisharpPackage : Output package depends on the input package and other input parameters like serverUrl", () => {
